requirements.txt
pandas
plotly
folium
streamlit-folium
seaborn